requirements.txt
SQLAlchemy==1.3.15
sqlalchemy-migrate==0.13.0
pymysql==0.9.3
sqlalchemy-utils==0.36.3